Avoid recursion in dask.core.get
Instead we use a topological sort

This cleans up tracebacks and profiling a bit when we have large complex
tasks or graphs (as we sometimes do when fusing many sequential tasks
together)
